fix: check for litee component, use file_tracking in handler
previous to this commit, when the handler fired, an initial call to
"file_tracking" autocmd was not made, and the filetree would open in a
collapsed state, despite being opened from a window.

additionally, the filetree code never checked if the current window is a
litee component before setting said window as its "invoking_win",
causing filetree jumps to happen within a litee component under the
right condition.

this commit now checks that a litee component window will never become
the filetree's "invoking_win".

Signed-off-by: ldelossa <louis.delos@gmail.com>