feat,popout: this commit wires in litee.nvim popout feature
this commit wires in the new litee.nvim popout feature.

the filetree can now be popped out from the panel to a floating window
with the command "LTPopOutFiletree"

additionally a config option "on_open" is added which takes the
arguments "popout" or "panel". if set to "panel" the first request for a
filetree is opened in the panel, same goes for "popout".